[TEST] Use nested openMAC timer and Tx/Rx interrupts
In order to handle Tx and Rx interrupts faster the priority is raised to
the highest. The timer interrupt is set to 2nd highest but can be
interrupted by the highest interrupt. This allows to free Tx and Rx
descriptors as fast as possible. In order to reduce the context switch
overhead two shadow registers are enabled in Nios II. The vectored
interrupt controller is used to allow interrupt stack with nested
interrupts.

Change-Id: I3c2f79484dda7d194716ac8160c88f55d32bd285