[TEMP] Sdoseq: Reset history buffer Tx fail flag on success
The history buffer Tx fail flag is reset if the frame is
forwarded to the kernel stack successfully. Still, the frame
could be dropped by the network interface or the network itself.
As an outcome the complete history buffer is *not* re-sent if a
new frame gets inserted. This optimizes utilization of the node's
asynchronous bandwidth.

Change-Id: I99b25dc0e4a6156a3211f7f364f2fe04fe7f2a18